- Refactor the useCumulative selector
