+++ jbook/src/components/code-cell.tsx
@@ -6,6 +6,7 @@ import Resizable from './resizable';
 import { Cell } from '../state';
 import { useActions } from '../hooks/use-actions';
 import { useTypedSelector } from '../hooks/use-typed-selector';
+import { useCumulativeCode } from '../hooks/use-cumulative-code';
 
 interface CodeCellProps {
   cell: Cell;
@@ -15,63 +16,23 @@ const CodeCell: React.FC<CodeCellProps> = (props) => {
   const { cell } = props;
   const { updateCell, createBundle } = useActions();
   const bundle = useTypedSelector(state => state.bundles[cell.id]);
-  const cumulativeCode = useTypedSelector((state) => {
-    const { data, order } = state.cells;
-    const orderedCells = order.map(id => data[id]);
-
-    const showFunc =       `
-    import _React from 'react';
-    import _ReactDOM from 'react-dom';
-
-    var show = (value) => {
-      const root = document.querySelector('#root');
-
-      if (typeof value === 'object') {
-        if(value.$$typeof && value.props) {
-          _ReactDOM.render(value, root);
-        } else {
-          root.innerHTML = JSON.stringify(value);
-        }
-      } else {
-        root.innerHTML = value;
-      }
-    };
-  `;
-  const showFuncNoOp = 'var show = () => {}';
-    const codes = [
-
-    ];
-    for(let c of orderedCells) {
-      if(c.type === 'code') {
-        if(c.id === cell.id) {
-          codes.push(showFunc);
-        } else {
-          codes.push(showFuncNoOp);
-        }
-        codes.push(c.content);
-      }
-      if(c.id === cell.id) break;
-    }
-    return codes;
-  });
-
-  // console.log(cumulativeCode);
+  const cumulativeCode = useCumulativeCode(cell.id);
 
   useEffect(() => {
     if(!bundle) {
-      createBundle(cell.id, cumulativeCode.join('\n'));
+      createBundle(cell.id, cumulativeCode);
       return;
     }
 
     const timer = setTimeout(async () => {
-      createBundle(cell.id, cumulativeCode.join('\n'));
+      createBundle(cell.id, cumulativeCode);
     }, 750);
 
     return () => {
       clearTimeout(timer);
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [cumulativeCode.join('\n'), cell.id, createBundle]);
+  }, [cumulativeCode, cell.id, createBundle]);
 
   return (
     <Resizable direction="vertical">
